requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
Django==4.2
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
msgpack==1.0.5
multidict==6.0.4
openai==0.27.4
python-dotenv==1.0.0
python-environ==0.4.54
pytz==2023.3
redis==4.5.4
requests==2.28.2
sqlparse==0.4.3
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
